+++ NoveltyProducer/Manager.py
@@ -49,9 +49,7 @@ class Manager:
         # init parameters
         self.verbose = verbose_
 
-    def create_pipeline(self, ip_, port_, topic_, frequency_, channel_name_, \
-        channel_limits_=defaults['channel_limits'], channel_frequency_=defaults['channel_frequency'], channel_type_=defaults['channel_type'],
-        pipeline_name_=defaults['pipeline_name'], dead_frequency_=defaults['dead_frequency'], dead_period_=defaults['dead_period']):
+    def create_pipeline(self, ip_, port_, topic_, frequency_, pipeline_name_=defaults['pipeline_name']):
         """Create pipeline and add each element to its dict. Start pipeline afterwards.
 
         Parameters:
@@ -59,55 +57,50 @@ class Manager:
         port_ (mandatory, int): Port of target host.
         topic_ (mandatory, string): Name of topic that data should be published on.
         frequency_ (mandatory, float): Frequency (in Hz) in that the data will be published on the given topic_.
-        channel_name_ (mandatory, string): Name of the new channel.
-        channel_limits_ (optional, list of floats): The lower/upper limits of the data.
-        channel_frequency_ (optional, float): Frequency (in Hz) in that the data will repeat itself.
-        channel_type_ (optional, string): The type of data production (e.g. default is sin-wave).
-        pipeline_name_ (optional, str): Name that a certain pipeline can later be associated with.
-        dead_frequency_ (optional, float): Frequency (in Hz) in that dead time will occur.
-        dead_period_ (optional, float): Duration (in seconds) of dead time every time it occurs.
         """
         # check inputs
-        self._check_pipeline_materials(ip_, port_, topic_, frequency_, channel_name_, channel_limits_, channel_frequency_, channel_type_, pipeline_name_, dead_frequency_, dead_period_)
+        self._check_inputs(ip_=ip_, port_=port_, topic_=topic_, frequency_=frequency_, pipeline_name_=pipeline_name_)
         # add host
         host_id = self._add_connection(ip_, port_)
         # add topic
         topic_id = self._add_topic(topic_, frequency_)
-        # add channel
-        channel_id = self._add_channel(channel_name_, channel_limits_, channel_frequency_, channel_type_, dead_frequency_, dead_period_)
         # add everything to pipeline
-        pipeline_id = self._add_pipeline(pipeline_name_, host_id, topic_id, channel_id)
+        pipeline_id = self._add_pipeline(pipeline_name_, host_id, topic_id)
         # init and assign generator for each new pipeline
         self._add_handlers(pipeline_id)
         # add new job in scheduler
-        self.Scheduler.add_job(func=self.publish_data, trigger='interval', seconds=(1/frequency_), id=str(pipeline_id), kwargs={'id_':pipeline_id})
+        self.Scheduler.add_job(func=self.publish_data, trigger='interval', seconds=(1/frequency_), id=str(pipeline_id), kwargs={'pid_':pipeline_id})
+        # pause job since no channel is yet on pipeline
+        self.Scheduler.get_job(str(pipeline_id)).pause()
         # return id of pipeline that has just been created
         return pipeline_id
 
-    def add_channel_to_pipeline(self, id_, name_, limits_=defaults['channel_limits'], frequency_=defaults['channel_frequency'], 
-                                type_=defaults['channel_type'], dead_frequency_=defaults['dead_frequency'], dead_period_=defaults['dead_period']):
-        """Add another channel to an already existing pipeline.
+    def add_function(self, pid_, channel_name_, limits_=defaults['channel_limits'], frequency_=defaults['channel_frequency'], 
+                     type_=defaults['channel_type'], dead_frequency_=defaults['dead_frequency'], dead_period_=defaults['dead_period']):
+        """Add a function to an already existing pipeline.
         
         Parameters:
-        id_ (mandatory, int): ID of pipeline.
-        name_ (mandatory, string): Name of channel.
+        pid_ (mandatory, int): ID of pipeline.
+        channel_name_ (mandatory, string): Name of channel.
         limits_ (optional, list of floats): The lower/upper limits of the data.
         frequency_ (optional, float): Frequency (in Hz) in that the data will repeat itself.
         """
+        # check inputs
+        self._check_inputs(channel_name_=channel_name_, channel_limits_=limits_, channel_frequency_=frequency_, channel_type_=type_, dead_frequency_=dead_frequency_, dead_period_=dead_period_)
         # add channel
-        cid = self._add_channel(name_=name_, limits_=limits_, frequency_=frequency_, type_=type_, dead_frequency_=dead_frequency_, dead_period_=dead_period_)
+        cid = self._add_channel(name_=channel_name_, limits_=limits_, frequency_=frequency_, type_=type_, dead_frequency_=dead_frequency_, dead_period_=dead_period_)
         # add channel to target pipeline
-        self.pipelines[id_]['channel_id'].append(cid)
+        self.pipelines[pid_]['channel_id'].append(cid)
         # is pipeline currently inactive?
-        if (self.pipelines[id_]['active'] == 0):
+        if (self.pipelines[pid_]['active'] == 0):
             # switch pipeline on
-            self.switch_pipeline(id_)
+            self.switch_pipeline(pid_)
         # get new generator and pass it to technican
-        self._update_technican(id_)
+        self._update_technican(pid_)
         # return id of channel that has just been added
         return cid
 
-    def remove_channel_from_pipeline(self, cid_):
+    def remove_channel(self, cid_):
         """Remove existing channel from each pipeline.
         
         Parameters:
@@ -128,17 +121,17 @@ class Manager:
                 # call corresponding technican.
                 self._update_technican(pid)
                 
-    def publish_data(self, id_):
+    def publish_data(self, pid_):
         """Get data and publish it to target host.
 
         Parameters:
-        id_ (mandatory, int): ID of pipeline.
+        pid_ (mandatory, int): ID of pipeline.
         """
         # get topic, channelname and data in json format
-        topic = self.topics[id_]['topic']
-        jdata = self.handlers[id_]['technican'].get_payload()
+        topic = self.topics[pid_]['topic']
+        jdata = self.handlers[pid_]['technican'].get_payload()
         # publish data via client
-        ret = self.handlers[id_]['mqtt'].publish(topic, jdata)
+        ret = self.handlers[pid_]['mqtt'].publish(topic, jdata)
         
         return ret
     
@@ -304,14 +297,13 @@ class Manager:
         # return the id that has just been added
         return id
 
-    def _add_pipeline(self, name_, host_id_, topic_id_, channel_id_):
+    def _add_pipeline(self, name_, host_id_, topic_id_):
         """Adding a new entry in the pipeline dictionary.
         
         Parameters:
         name_ (mandatory, str): Name of the new pipeline.
         host_id_ (mandatory, int): ID of host in host dictionary.
         topic_id_ (mandatory, int): ID of topic in topic dictionary.
-        channel_id_ (mandatory, int): ID of channel in channel dictionary.
 
         Note:
         - name_ can also be None or an empty string.
@@ -333,8 +325,8 @@ class Manager:
             'name':name,
             'host_id':host_id_,
             'topic_id':topic_id_,
-            'channel_id':[channel_id_],
-            'active':1
+            'channel_id':[],
+            'active':0
         }
         # return the id that has just been added
         return id
@@ -392,60 +384,71 @@ class Manager:
         
         return name
         
-    def _check_pipeline_materials(self, ip_, port_, topic_, frequency_, channel_name_, channel_limits_, channel_frequency_, channel_type_,  pipeline_name_, dead_frequency_, dead_period_):
+    def _check_inputs(self, **kwargs):
         """Check all inputs that are used to create a pipeline"""
         # ip_
-        if not isinstance(ip_, str):
-            raise InvalidInputTypeError("Content of ip_ is type %s but should be a of type string." % type(ip_))
-            
+        if 'ip_' in kwargs:
+            if not isinstance(kwargs['ip_'], str):
+                raise InvalidInputTypeError("Content of ip_ is type %s but should be a of type string." % type(kwargs['ip_']))
+
         # port_
-        if not isinstance(port_, int):
-            raise InvalidInputTypeError("Content of port_ is type %s but should be a of type int." % type(port_))
-        if ((port_ < 0) | (port_ > 65535)):
-            raise InvalidInputValueError("Value of port_ (%s) is not in valid port range (0 - 65535)." % str(port_))
+        if 'port_' in kwargs:
+            if not isinstance(kwargs['port_'], int):
+                raise InvalidInputTypeError("Content of port_ is type %s but should be a of type int." % type(kwargs['port_']))
+            if ((kwargs['port_'] < 0) | (kwargs['port_'] > 65535)):
+                raise InvalidInputValueError("Value of port_ (%s) is not in valid port range (0 - 65535)." % str(kwargs['port_']))
 
         # topic_
-        if not isinstance(topic_, str):
-            raise InvalidInputTypeError("Content of topic_ is type %s but should be a of type string." % type(topic_))
-            
+        if 'topic_' in kwargs:
+            if not isinstance(kwargs['topic_'], str):
+                raise InvalidInputTypeError("Content of topic_ is type %s but should be a of type string." % type(kwargs['topic_']))
+
         # frequency_
-        if not isinstance(frequency_, (int, float)):
-            raise InvalidInputTypeError("Content of frequency_ is type %s but should be a of type int or float." % type(frequency_))
-        if frequency_ <= 0:
-            raise InvalidInputValueError("Value of frequency_ (%s [Hz]) is negative or zero but should be positive." % str(frequency_))
+        if 'frequency_' in kwargs:
+            if not isinstance(kwargs['frequency_'], (int, float)):
+                raise InvalidInputTypeError("Content of frequency_ is type %s but should be a of type int or float." % type(kwargs['frequency_']))
+            if kwargs['frequency_'] <= 0:
+                raise InvalidInputValueError("Value of frequency_ (%s [Hz]) is negative or zero but should be positive." % str(kwargs['frequency_']))
             
         # channel_name_
-        if not isinstance(channel_name_, str):
-            raise InvalidInputTypeError("Content of channel_name_ is type %s but should be a of type string." % type(channel_name_))
+        if 'channel_name_' in kwargs:
+            if not isinstance(kwargs['channel_name_'], str):
+                raise InvalidInputTypeError("Content of channel_name_ is type %s but should be a of type string." % type(kwargs['channel_name_']))
             
         # channel_limits_
-        if not isinstance(channel_limits_, list):
-            raise InvalidInputTypeError("Content of channel_limits_ is type %s but should be a of type list." % type(channel_limits_))
-        if not all(isinstance(x, (int, float)) for x in channel_limits_):
-            raise InvalidInputValueError("Not all values of channel_limits_ are of type int or float.")
+        if 'channel_limits_' in kwargs:
+            if not isinstance(kwargs['channel_limits_'], list):
+                raise InvalidInputTypeError("Content of channel_limits_ is type %s but should be a of type list." % type(kwargs['channel_limits_']))
+            if not all(isinstance(x, (int, float)) for x in kwargs['channel_limits_']):
+                raise InvalidInputValueError("Not all values of channel_limits_ are of type int or float.")
             
         # channel_frequency_
-        if not isinstance(channel_frequency_, (int, float)):
-            raise InvalidInputTypeError("Content of channel_frequency_ is type %s but should be a of type int or float." % type(channel_frequency_))
-        if channel_frequency_ <= 0:
-            raise InvalidInputValueError("Value of channel_frequency_ (%s [Hz]) is negative or zero but should be positive." % str(channel_frequency_))
+        if 'channel_frequency_' in kwargs:
+            if not isinstance(kwargs['channel_frequency_'], (int, float)):
+                raise InvalidInputTypeError("Content of channel_frequency_ is type %s but should be a of type int or float." % type(kwargs['channel_frequency_']))
+            if kwargs['channel_frequency_'] <= 0:
+                raise InvalidInputValueError("Value of channel_frequency_ (%s [Hz]) is negative or zero but should be positive." % str(kwargs['channel_frequency_']))
             
         # channel_type_
-        if not isinstance(channel_type_, str):
-            raise InvalidInputTypeError("Content of channel_type_ is type %s but should be a of type string." % type(channel_type_))
+        if 'channel_type_' in kwargs:
+            if not isinstance(kwargs['channel_type_'], str):
+                raise InvalidInputTypeError("Content of channel_type_ is type %s but should be a of type string." % type(kwargs['channel_type_']))
             
         # pipeline_name_
-        if not isinstance(pipeline_name_, str):
-            raise InvalidInputTypeError("Content of pipeline_name_ is type %s but should be a of type string." % type(pipeline_name_))
+        if 'pipeline_name_' in kwargs:
+            if not isinstance(kwargs['pipeline_name_'], str):
+                raise InvalidInputTypeError("Content of pipeline_name_ is type %s but should be a of type string." % type(kwargs['pipeline_name_']))
             
         # dead_frequency_
-        if not isinstance(dead_frequency_, (int, float)):
-            raise InvalidInputTypeError("Content of dead_frequency_ is type %s but should be a of type int or float." % type(dead_frequency_))
-        if dead_frequency_ <= 0:
-            raise InvalidInputValueError("Value of dead_frequency_ (%s [Hz]) is negative or zero but should be positive." % str(dead_frequency_))
+        if 'dead_frequency_' in kwargs:
+            if not isinstance(kwargs['dead_frequency_'], (int, float)):
+                raise InvalidInputTypeError("Content of dead_frequency_ is type %s but should be a of type int or float." % type(kwargs['dead_frequency_']))
+            if kwargs['dead_frequency_'] <= 0:
+                raise InvalidInputValueError("Value of dead_frequency_ (%s [Hz]) is negative or zero but should be positive." % str(kwargs['dead_frequency_']))
             
         # dead_period_
-        if not isinstance(dead_period_, (int, float)):
-            raise InvalidInputTypeError("Content of dead_period_ is type %s but should be a of type int or float." % type(dead_period_))
-        if dead_period_ < 0:
-            raise InvalidInputValueError("Value of dead_period_ (%s [Hz]) is negative but should be zero or positive." % str(dead_period_))
+        if 'dead_period_' in kwargs:
+            if not isinstance(kwargs['dead_period_'], (int, float)):
+                raise InvalidInputTypeError("Content of dead_period_ is type %s but should be a of type int or float." % type(kwargs['dead_period_']))
+            if kwargs['dead_period_'] < 0:
+                raise InvalidInputValueError("Value of dead_period_ (%s [Hz]) is negative but should be zero or positive." % str(kwargs['dead_period_']))